server/requirements.txt
SQLAlchemy~=2.0.23
boto3~=1.29.6
botocore~=1.32.6
fastapi~=0.104.1
starlette~=0.27.0

psycopg2-binary
python-multipart
openai-whisper
uvicorn~=0.24.0.post1

accelerate==0.21.0
bitsandbytes==0.40.2
peft==0.5.0
transformers==4.34.0
sentencepiece
torch~=2.1.1
tqdm~=4.66.1
scipy

python-dotenv~=1.0.0
